Created sparseMatrixExtended and sparseMatrixExtendedTools
sparseMatrixExtended stores sparse matrices with every coefficient as a 4x4 RectangularMatrix.

sparsematrixExtendedTools solves the linear system using PETSc. The matrix coefficients are 4x4 tensors and the source is a 4x1 vector for each row in the matrix.

The enforceFixedDof function still needs to be changed to deal with displacement and pressure (TO DO).