Integrate explicitSolidDynamics toolkit  as explicitGodunovCCSolid solidModel
This is a complete integration of the cell centered explicitSolidDynamics toolkit  as a solidModel subclass named explicitGodunovCCSolid.
However, explicitSolidDynamics solves for linear momentum and deformation gradient fields while solids4Foam solves for displacement field.
The next step is to find relations between the fields, and replace the dependencies on the original explicitSolidDynamics libraries with the ones that are available in solids4Foam.

cantilever beam bending test case is added to test the results with solids4Foam with the new solid model.